Introduce ability to pass ZOOKEEPER_PORTS
It is to allow to run more than 1 zookeeper cluster on the same nodes, e.g. one for config and another one fo analytics.
This parameter if specified causes the script to make ZOO_SERVER list for inherited docker-entrypoint.
If ZOO_SERVER is defined then it is passed as is. It is to keep inherited container interface.