Add support for CentOS mirrors in external/kafka
In pursuit of minimizing the external connectivity requirements, added
the possibility to configure yum mirrors in kafka container. This
is probably the last place that needs the mirror configuration, as other
extarnal containers don't install additional packages during build.

There is no good solution to the situation when no additional repos are
required. If no *.repo files are present in the build context,
the build will fail. Kafka doesn't need the contrail repo, so there is
no point in installing the contrail.repo file here. To make the default
build work, added the empty.repo file.

Ultimately, we should package the kafka release into rpm, add it to
the contrail third party repo and install this repo config file here,
thus adding a functional file that is present also for the default build.

Change-Id: I046a1de08049a440a8f5e9844592fbaad33ba490