For introduce net-watchdog for dpdk case.
This container periodically checks that vhost0
is up and if it is check that it has IP,
if there is no IP then it assign it.
This is a qorkaround for now because
vhost0 can dissappear any time if dpdk container
restarts.
For non-dpdk case the vhost0 init actions are done
in agent container.

Change-Id: I5051b7c1cf123aee0f2b875be1e6c59b5c736486
Partial-Bug: #1743610