Introduce network-init container.
This container is to be used in non-dpdk
case for vhost0 initialization.
The call if init_vhost0 will be removed
from in agent container as soon as
all yamls be updated to use this container.

This is needed to be possible to re-use agent
container in orchestrators that manage
networks by themselves, like RDO.

Change-Id: I8a44d08454685de37425ac613675222d1ab43b56
Partial-Bug: #1743610