Fixes for OSP
ifcfg-vhost0 is prepared by os-net-config and
ifup vhost0 is called by os-net-config as well.
This is done in OSP at the early stage of deployment
before any ..NODES are set => running agent container
is not an option for vhost0 initializatio. The only way is kept
is to allow to ifup to create vhost device by itself.

Depends-On: I41b8c2292004bcce6f00d356f2e8fc6d1ade3396
Change-Id: I9251e607396d0a81e91cf44820dffa48d8be4361
Partial-Bug: #1770137