Force installation of python-docker-py
This package is obsoleted by python2-docker that is
available in queens repos. So, in queens build
nodemgr and contrail-status containers cant work
because they use old API.
This is temporary fix. After contrail-status and nodemgr
be switched to new API python-docker-py is to be replaced
with python2-docker.

Change-Id: I828d315f3d50ee4f284cb6ab999932a7c5ae738e
Partial-Bug: #1770137