Closes-Bug: #1783567
Getting the details of the VM from the DMI table of the provider.
In case of Axure OR Google, vhost0 needs to also get the DHCP address
which will be the same as that of the physical interface it is
Xconnected to.

This fix will enable proper vhost0 initialization in Azure and Google
clouds.

Using chassis_vendor instead of dmidecode

Change-Id: I742aaa3960133884f9652c1a802518f1b60b92e3