Add iptable config support to node-init container
This commit adds the ability to configure iptable rules on a node
through the node-init container.

This scripts works off the following environment variables to see
if iptable configuration is required. If required, then it deciphers
the table, chain and port info using addition env variables.

CONFIGURE_IPTABLES  -> Set to true/yes/enabled if iptable configuration is required.
                       Default value is 'false'
IPTABLES_OPEN_PORTS -> Ports to be opened up in iptables, delimited by ','.
                       Multiple ports or port-ranges may be specified.
                       Default value is ''
IPT_CHAIN_NAME      -> Name of the chain in iptables.
                       This chain, if specified, MUST exist.
                       Default value is "INPUT"
IPT_TABLE_NAME      -> Name of the table in iptables.
                       This table, if specified, MUST exist.
                       Default value is "filter"

With this change, ports for a kubernetes pod can be opened using the node-init
container as pod's pre-init container.

This negates the need to know/use host's firewall mechanism(i.e firewalld, ipables etc).

Change-Id: Iff634d3596b9d4bc07385069cadd53a7363abcf4
Partial-Bug: #1795750