Added optional env file for vhost ifup/ifdown
- it is for using during ff upgrade:
in osp10 there is old os-net-config version
that during ffupgrade prepare command performs
reinit of vhost0 interface. At this moment vrouter is
already upgraded to containerized Contrail
(because it requries workload migration and reboots).
But because os-net-config is outdated it doesnt provided
required options in ifcfg-vhost0 that breaks vhost0
reinit at this stage. These files are to be created
manually at this stage of upgrade with variables
that missed by old os-net-config.

Change-Id: Iba518e72daa1dbd394298ec21bd65a31b1af0a35
Partial-Bug: #1801956