Update patch set 4
Patch Set 4:

> Patch Set 3: Code-Review-1
> 
> clean up in base containers leads to increasing time build. base containers must not be pushed to dockerhub and their size shouldn't not be matter.

Interesting,  It seems that from the report below, it looks like the build of this change set is about the same as build 24 of R1908.  See http://logs.opencontrail.org/periodic-nightly/review.opencontrail.org/R1908/24/.

Build 24:
* Centos 7.4 Queens: 1hr 13m
* RHEL 7 Queens: 1hr 23m

Merge Request:
* Centos 7.4 Queens: 48m   -25m   -34%
* RHEL 7 Queens: 1hr 42m   +19m   +23%

Note that this is not a huge sample set, but I strongly suspect that since the change is coincident with such a wide range of variability, that the variability has little to do with the change.  Generally it seems like noise.

Regarding the size, when a container is based on another one, all the layers are inherited.  If one layer adds 2GB of files and the next layer delete most of it, the 2GB layer is still present.  This is why it's best practice to clean up when installing packages with yum in a single layer.  See https://developers.redhat.com/blog/2016/03/09/more-about-docker-images-size/.

Patch-set: 4
Reviewer: Gerrit User 822 <822@c72cd27d-1743-4a55-84a4-cd13b8400c52>