Update patch set 2
Patch Set 2:

Guys, it would be good if can describe the feature or link a bug with description as early as possible. in this case we can discuss it and make an agreement how to proceed. in other case it can be a situation when some work was done but it can't fit into current product.

Patch-set: 2
Reviewer: Gerrit User 382 <382@c72cd27d-1743-4a55-84a4-cd13b8400c52>