closes-jira-bug: cem-9292
In deployments where there is already an existing
L3 interface with NOARP flag, creating another
interface and enabling it will lead to incorrect
handling of the pkts as was the case reported in
this issue where Calico had initialized tunl0
interface.

VTI interface as such need not be enabled as the
associated interface is enabled. In this case, the
decrypt0 interface.

Change-Id: Ife85f02635a2f245d9b52b029f2b551413ebe822