closes-jira-bug: CEM-9507
Change-Id: I176f40650654d3cb71faabd850d6b61b7e98d570

In case of DHCP, vhost0 takes over the DHCP and gets set with the
MTU as provided by the DHCP interface MTU option. However the physical
goes to the default MTU as it is not running DHCP anymore.
To avoid descrepancies that leads to a lot of issues, it is best to
ensure the physical interface is also set to the same MTU as it was.

Change-Id: Ic0e3f0ea1af0e2d2d8b4bc512f444398b71bff0f