closes-jira-bug: CEM-9507
Change-Id: Ia6bc2234d97244178bf1c3798680afe6d51f8c26

In case of DHCP, vhost0 takes over the DHCP and gets set with the
MTU as provided by the DHCP interface MTU option. However the physical
goes to the default MTU as it is not running DHCP anymore.
To avoid descrepancies that leads to a lot of issues, it is best to
ensure the physical interface is also set to the same MTU as it was.

Change-Id: Id4aada301af5304a97725ecc6f5df45d234c0140