closes-jira-bug: CEM-9599
Change-Id: Id02c4cdca91df725582fee9f704b5a404f1f4503

Root Cause: Since the dhclient run on vhost0 was providing
the offers for all the interfaces including the one that
has the default route (routers option), which causes the
route for default be updated with vhost0. The ways it is
setup on contrail-multicloud-GW in Azure, Google and AWS
is that the default is always on the management and
control/data interface in the second/private interface.
Since the default route got updated with the vhost0 interface,
the system could not reach out.

Fix: The fix ensures the DHCP request is made only for the vhost0
interface. This addresses the case where vhost0 is connected to
the physical interface that could or could not have the default
route.

Change-Id: I6c32cfd062253387e9b5caf62bfae4ef7e6cf28a