closes-jira-bug: CEM-9599
Change-Id: I9af9b1c323e2eab36c3120fa668fea68688e3a07

Root Cause: Since the dhclient run on vhost0 was providing
the offers for all the interfaces including the one that
has the default route (routers option), which causesthe
route for default be updated with vhost0. The ways it is
setup on contrail-multicloud-GW in Azure, Google and AWS
is that the default is always on the management and
control/data interface in the second/private interface.
Since the default route got updated with the vhost0 interface,
the system could not reach out.

Fix: The fix ensures the DHCP request is made only for the vhost0
interface. This addresses the case where vhost0 is connected to
the physical interface that could or could not have the default.

Change-Id: I782aaec5592895c5be5e112bedada5929e89c78b