closes-jira-bug: CEM-9599
Change-Id: I9af9b1c323e2eab36c3120fa668fea68688e3a07

Root Cause: Since the dhclient run on vhost0 was providing
the details on the routers option with the default GW, the
route for default was updated with vhost0. In a multi-interface
setup, the default is always on the management and control/data
interface in the second/private interface. Since the default route
got updated with the vhost0 interface, the system cannot reach out.

Fix: The fix ensures the DHCP request is made only for the vhost0
interface and hence if the physical interface does not have the
default GW, then the DHCP router option will not be set and the
default route will not be updated with vhost0.

Change-Id: I782aaec5592895c5be5e112bedada5929e89c78b