closes-jira-bug: cem-9847
Root-Cause: Contrail-vrouter-agent container manages the plumbing of
vhost0 and copies a script to be managed by ifup. For multicloud,
deployments, the model is to have most of the interface management
within the agent container without doing much on the host.

Fix: The fix will ensure consistency in management vhost0 interface
across multi-cloud envs. In future when vhost0 is managed by
NetworkManager, it will be easy to use NetworkManager across all
multi-cloud envs.

Change-Id: Idb66ddc0cf7fb511bece25d83d9cfc9104d5077f