Update patch set 1
Patch Set 1:

We have done some tests with it and it seems, while this change ffixes nameserver insdeed, it shouldn't be merged yet.

The reason is, with this change, on MC gateways the default gateway is set to go via vhost0. This means the GWs are cut off as soon as DHCP response arrives and is processed.

Do you have any ideas on fixing it? We could in theory ignore default route, but in that case, we'd cut off Compute Nodes instead.

Patch-set: 1
CC: Gerrit User 532 <532@c72cd27d-1743-4a55-84a4-cd13b8400c52>