Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> We have done some tests with it and it seems, while this change ffixes nameserver insdeed, it shouldn't be merged yet.
> 
> The reason is, with this change, on MC gateways the default gateway is set to go via vhost0. This means the GWs are cut off as soon as DHCP response arrives and is processed.
> 
> Do you have any ideas on fixing it? We could in theory ignore default route, but in that case, we'd cut off Compute Nodes instead.

This change will not affect the default route on the gateway. This was already working for Azure and Google. The only change is that, AWS is also included. The dhclient is also only applied for the vhost0. So if vhost0 does not have default gateway, then it should not update the default route. This is how it was working in Azure and Google since 5.0.

Patch-set: 1
Reviewer: Gerrit User 53 <53@c72cd27d-1743-4a55-84a4-cd13b8400c52>