Update patch set 3
Patch Set 3:

Long term, all references to "contrail" has to be removed from the codebase as that is a Juniper trademark and this code is technically owned by the Linux Foundation now.  Juniper signed legal agreements to that effect.

That being said, making these changes in any kind of expeditious manner is clearly impossible.  It will take a while to get it done and, in fact, it may never be completely done, but we have to take a run at it.

So, whenever possible, we should be making these changes starting with the most visible, such as repo names, then next most visible (file names), and things that are least visible (function names) can stay as they are for quite a while and just be changed when re-factoring code.

Patch-set: 3
CC: Gerrit User 713 <713@c72cd27d-1743-4a55-84a4-cd13b8400c52>