Update patch set 2
Patch Set 2: Code-Review-1

Damian, let's not capitalize the names of new variables, since they are not externally exposed parameters but rather internal ones. CAPITALIZED are only the env parameters.
And please add the comment, explaining that it's a temporary workaround and how it is supposed to be reworked in the future

Patch-set: 2
Reviewer: Gerrit User 445 <445@c72cd27d-1743-4a55-84a4-cd13b8400c52>
Label: Code-Review=-1