Update patch set 2
Patch Set 2:

Here is the idea:

When agent container starts up, there is the ifup-vhost that is being copied to the host and after determining that it can be used to up the vhost0 interface the call is made to "ifup vhost0". 

This works for most cases but not for case where the DHCP lease is set to a short time. vrouter does not have the support for natively supporting DHCP on the fabric xconnect interface (vhost0) with all the options. In the interim, this fix, will check ensure that the lease is superseded with value set to expire in 5 years, especially in the public clouds viz., AWS, Azure, and Google. 

The other possibility is to make change in ifup-vhost such that it support -cf arg and change docker-compose.yaml to mount /etc/dhcp so that the conf can be read by the container. 

I personally like the option where the dhclient's lifetime is linked to the agent container.

Patch-set: 2
Reviewer: Gerrit User 53 <53@c72cd27d-1743-4a55-84a4-cd13b8400c52>