support docker-less context by contrail-status script
redhat tends to replace docker with podman utility thus the status
script may not find docker tools around.
let's check if the tool exists before executing container logic.
it is better to use podman as well even in case of a latest
kubernates installation because the cri-o backend doesn't have
required functionality, fi it always needs a pod to create a
container even if a short-living one. still the code inside the
container needs to cooperate with cri-o if the later is in charge
of containers locally. the solution is to detect a kubelet process
for the switch between podman only and kubernetes branches.

Change-Id: I44f07ef9ab7ffa5ec4f83e0e437f82b731c9f96d